Us Univ Automation/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
jinja2